Seems to train fine but slow
