+++ conf/models/segmentation.yaml
@@ -320,7 +320,7 @@ models:
         type: pointnet2_customkernel
         down_conv:
             module_name: SADenseModule
-            ratios: [0.25, 0.25, 0.25, 0.25]
+            ratios: [0.33, 0.33, 0.33, 0.33]
             radius: [[0.05, 0.1], [0.1, 0.2], [0.2, 0.4], [0.4, 0.8]]
             down_conv_nn: [[FEAT + 3, 32, 32, 64], 
                             [64 * 2 + 3, 64, 96, 128],
@@ -338,6 +338,24 @@ models:
             module_name: GlobalDenseBaseModule
             aggr: max
             nn: [256 + 3, 256, 512, 1024]
+        mlp_cls: 
+            nn: [128, 128, 128]
+            dropout: 0.5
+
+    pointnet2_simple:
+        type: pointnet2_customkernel
+        down_conv:
+            module_name: SADenseModule
+            ratios: [0.3, 0.3]
+            radius: [[0.1, 0.2], [0.2, 0.4]]
+            down_conv_nn: [[FEAT + 3, 32, 64, 128], 
+                            [128 * 2 + 3, 256, 384, 512]]
+            radius_num_points: [[16, 32], [16, 32]]
+        up_conv:
+            module_name: DenseFPModule
+            up_conv_nn: [[512 * 2 + 128 * 2, 512, 256], 
+                            [256 + FEAT, 128, 128, 128]]
+            skip: True 
         mlp_cls: 
             nn: [128, 128, 128]
             dropout: 0.5
